Add files via upload
This is an exploratory analysis of a data set about poker. I made this during my minor (ERP & BI) after having completed an introductory course to R on Datacamp. ggplot2 and corrplot are used, along with basic R.  Because of unfamiliarity with statistical techniques at the time, conclusions are mostly 'ad-hoc'.